+++ GFramework.Godot/Architectures/IGodotModule.cs
@@ -1,7 +1,8 @@
 ﻿using GFramework.Core.Abstractions.Architectures;
+using GFramework.Core.Architectures;
 using Godot;
 
-namespace GFramework.Godot.Architecture;
+namespace GFramework.Godot.Architectures;
 
 /// <summary>
 ///     Godot模块接口，定义了Godot引擎中模块的基本行为和属性
@@ -17,7 +18,7 @@ public interface IGodotModule : IArchitectureModule
     ///     当模块被附加到架构时调用
     /// </summary>
     /// <param name="architecture">要附加到的架构实例</param>
-    void OnAttach(Core.Architectures.Architecture architecture);
+    void OnAttach(Architecture architecture);
 
     /// <summary>
     ///     当模块从架构分离时调用
